# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Tuneverse Writing API",
    "description": "API for Tuneverse Writing"
  },
...
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/api/user/subscriptions": {
      "get": {
        "tags": ["Subscription"],
        "summary": "Get all subscriptions",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/subscriptions": {
      "post": {
        "tags": ["Subscription"],
        "summary": "Create subscription",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Subscription object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "logo": {
                  "type": "string"
                },
                "subscriptionTypeId": {
                  "type": "integer"
                },
                "countryId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created subscription"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/subscriptions/{id}": {
      "get": {
        "tags": ["Subscription"],
        "summary": "Get subscription by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/subscriptions/{id}": {
      "put": {
        "tags": ["Subscription"],
        "summary": "Change subscription by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "logo": {
                  "type": "string"
                },
                "subscriptionTypeId": {
                  "type": "integer"
                },
                "countryId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptions/{id}": {
      "delete": {
        "tags": ["Subscription"],
        "summary": "Delete subscription by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptions/{id}/logicalDelete": {
      "patch": {
        "tags": ["Subscription"],
        "summary": "Logical delete subscription by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptions/{id}/logicalRestore": {
      "patch": {
        "tags": ["Subscription"],
        "summary": "Logical restore subscription by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/subscriptionTypes": {
      "get": {
        "tags": ["SubscriptionType"],
        "summary": "Get all subscriptionTypes",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/subscriptionTypes": {
      "post": {
        "tags": ["SubscriptionType"],
        "summary": "Create subscriptionType",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "SubscriptionType object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "cost": {
                  "type": "double"
                },
                "duration": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created subscriptionType"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/subscriptionTypes/{id}": {
      "get": {
        "tags": ["SubscriptionType"],
        "summary": "Get subscriptionType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "SubscriptionType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/subscriptionTypes/{id}": {
      "put": {
        "tags": ["SubscriptionType"],
        "summary": "Change subscriptionType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "SubscriptionType ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "cost": {
                  "type": "double"
                },
                "duration": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptionTypes/{id}": {
      "delete": {
        "tags": ["SubscriptionType"],
        "summary": "Delete subscriptionType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "SubscriptionType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptionTypes/{id}/logicalDelete": {
      "patch": {
        "tags": ["SubscriptionType"],
        "summary": "Logical delete subscriptionType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "SubscriptionType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/subscriptionTypes/{id}/logicalRestore": {
      "patch": {
        "tags": ["SubscriptionType"],
        "summary": "Logical restore subscriptionType by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "SubscriptionType ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/user/taskComments": {
      "get": {
        "tags": ["TaskComment"],
        "summary": "Get all taskComments",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["TaskComment"],
        "summary": "Create taskComment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "TaskComment object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                },
                "taskId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created taskComment"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/user/taskComments/{id}": {
      "get": {
        "tags": ["TaskComment"],
        "summary": "Get taskComment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TaskComment ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/moderator/taskComments/{id}": {
      "put": {
        "tags": ["TaskComment"],
        "summary": "Change taskComment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TaskComment ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string"
                },
                "taskId": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/taskComments/{id}": {
      "delete": {
        "tags": ["TaskComment"],
        "summary": "Delete taskComment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TaskComment ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/taskComments/{id}/logicalDelete": {
      "patch": {
        "tags": ["TaskComment"],
        "summary": "Logical delete taskComment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TaskComment ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/taskComments/{id}/logicalRestore": {
      "patch": {
        "tags": ["TaskComment"],
        "summary": "Logical restore taskComment by id",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "TaskComment ID",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}